delete UploadFileInput, use FileInput everywhere instead
