+++ frontend/components/inputs/FileInput.js
@@ -2,17 +2,16 @@ import React, {useRef, useState} from 'react';
 import PropTypes from "prop-types";
 
 /**
- * A component for file input with a specified callback onSubmit. An alternative to
- * UploadFileInput allowing for different handling of file inputs besides direct uploading to an
- * api endpoint.
+ * A component for file input with a specified callback onSubmit (usually an API request).
  */
 const FileInput = ({onSubmit}) => {
-    const [submitted, setSubmitted] = useState({"file": false});
+    const [submitted, setSubmitted] = useState(false);
     const [tempFile, setTempFile] = useState(null);
-    const fileRef = useRef(null);
+    const fileInputRef = useRef(null);
+
     const handleSubmitFile = (event) => {
         event.preventDefault();
-        setSubmitted(prevSubmitted => ({...prevSubmitted, "file": true}));
+        setSubmitted(true);
         onSubmit(tempFile);
     };
 
@@ -21,36 +20,39 @@ const FileInput = ({onSubmit}) => {
     };
 
     const clearFile = () => {
-        const file = fileRef.current;
+        const file = fileInputRef.current;
         file.value = null;
         setTempFile(null);
-        setSubmitted(prevSubmitted => ({...prevSubmitted, "file": false}));
+        setSubmitted(false);
     };
 
     return (
-        <>
-            <form className="form-inline">
-                <div className="form-group">
-                    <input className="form-control my-3" type="file" ref={fileRef}
-                        accept="text/csv" disabled={submitted.file}
-                        onChange={handleFileInput}/>
-                    {tempFile &&
-                    <button className="btn btn-secondary ml-1"
-                        type="button"
-                        onClick={clearFile} disabled={tempFile === null}
-                        data-dismiss="fileupload">Clear</button>
-                    }
-                </div>
+        <form className="form-inline">
+            <div className="form-group">
+                <input
+                    className="form-control my-3"
+                    type="file"
+                    ref={fileInputRef}
+                    accept="text/csv"
+                    disabled={submitted}
+                    onChange={handleFileInput}
+                />
                 {tempFile &&
-                <button id="upload-file" className="btn btn-primary ml-1"
-                    disabled={tempFile === null}
+                <button
+                    className="btn btn-secondary ml-1" type="button"
+                    onClick={clearFile}
+                >Clear</button>
+                }
+                {tempFile && !submitted &&
+                <button
+                    id="upload-file" className="btn btn-primary ml-1"
                     type={"submit"}
                     onClick={handleSubmitFile}>
                     Upload file
                 </button>
                 }
-            </form>
-        </>
+            </div>
+        </form>
     );
 };
